IO: Make stream size and pos i64 (#3069)

+++ Source/Urho3D/IO/NamedPipe.h
@@ -28,9 +28,9 @@ public:
     ~NamedPipe() override;
 
     /// Read bytes from the pipe without blocking if there is less data available. Return number of bytes actually read.
-    unsigned Read(void* dest, unsigned size) override;
+    i32 Read(void* dest, i32 size) override;
     /// Set position. No-op for pipes.
-    unsigned Seek(unsigned position) override;
+    i64 Seek(i64 position) override;
     /// Write bytes to the pipe. Return number of bytes actually written.
     i32 Write(const void* data, i32 size) override;
     /// Return whether pipe has no data available.
